+++ src/effects/feats/squireOfSolamnia.js
@@ -1,0 +1,40 @@
+import { baseItemEffect } from "../effects.js";
+import { loadMacroFile, generateItemMacroFlag, generateMacroChange } from "../macros.js";
+
+export async function squireOfSolamniaEffect(document) {
+  let effect = baseItemEffect(document, document.name);
+  const itemMacroText = await loadMacroFile("feat", "squireOfSolamnia.js");
+  document.flags["itemacro"] = generateItemMacroFlag(document, itemMacroText);
+  effect.changes.push(generateMacroChange(`"${document.name}"`));
+  effect.transfer = false;
+
+  effect.changes.push(
+    {
+      key: "flags.midi-qol.advantage.attack.mwak",
+      mode: CONST.ACTIVE_EFFECT_MODES.ADD,
+      value: "1",
+      priority: "20",
+    },
+    {
+      key: "flags.midi-qol.advantage.attack.rwak",
+      mode: CONST.ACTIVE_EFFECT_MODES.ADD,
+      value: "1",
+      priority: "20",
+    },
+    {
+      key: "flags.dnd5e.DamageBonusMacro",
+      value: "ItemMacro",
+      mode: CONST.ACTIVE_EFFECT_MODES.CUSTOM,
+      priority: 20,
+    }
+  );
+  setProperty(effect, "flags.dae.specialDuration", ["1Attack"]);
+  setProperty(effect, "flags.dae.selfTarget", true);
+  setProperty(effect, "flags.dae.selfTargetAlways", true);
+  document.effects.push(effect);
+
+  document.system.damage.parts = [];
+  document.system.actionType = null;
+
+  return document;
+}